add practice function for pointers and remove delete of player pointers
We add a function where we play around with some fundamental pointer
types.

We remove the delete of player pointers from `combat class` since that
class is not the owner of the players, it is just borrowing them.
Therefore we delete the players after they are done in the combat.